Fix stack use after return from pointers into a SmallVec
A SmallVec's contents may be inline, or behind an allocation. If they are
local, they are moved to a new location when the SmallVec is returned.
This means that the two pointers returned by aligned_packing_vec would
be dangling if the returned SmallVec were ever in he inline state.

This can be fixed by returning offsets instead of pointers, then
computing the pointers after the SmallVec is returned.